pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "Advent of Code solutions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "advent-of-code-data==2.0.4",
    "aocd==0.1",
    "aocd-example-parser==2023.12.20",
    "beautifulsoup4==4.12.3",
    "click==8.1.7",
    "intervaltree==3.1.0",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "Pebble==5.1.0",
    "Pygments==2.18.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "rich==13.9.4",
    "six==1.16.0",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "tzdata==2024.2",
    "urllib3==2.2.3",
]

[project.scripts]
aoc-run = "main:run"

[tool.setuptools]
py-modules = ["main", "utils"]

[tool.hatch.build.targets.wheel]
packages = ["."]
